BackImage/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna
mangum==0.17.0
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pydantic==2.1.1
pydantic_core==2.4.0
pytesseract==0.3.10
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
yolov5==7.0.12